requirements.txt
fastapi
uvicorn
pillow
numpy
torch==2.3.1
torchvision==0.18.1
facenet-pytorch
transformers
requests
gdown
python-multipart
pillow-heif==0.7.0
aiofiles